stage 1

+++ Common/cpp/Registries/WorkletsCache.cpp
@@ -18,8 +18,8 @@ jsi::Function function(jsi::Runtime &rt, const std::string& code) {
   return eval(rt, ("(" + code + ")").c_str()).getObject(rt).getFunction(rt);
 }
 
-std::shared_ptr<jsi::Function> WorkletsCache::getFunction(jsi::Runtime & rt, std::shared_ptr<FrozenObject> frozenObj) {
-  int workletId = frozenObj->map["__workletID"]->numberValue;
+std::shared_ptr<jsi::Function> WorkletsCache::getFunction(jsi::Runtime &rt, std::shared_ptr<FrozenObject> frozenObj) {
+  long long workletId = frozenObj->map["__workletID"]->numberValue;
   if (worklets.count(workletId) == 0) {
     jsi::Function fun = function(rt, frozenObj->map["asString"]->stringValue);
     std::shared_ptr<jsi::Function> funPtr(new jsi::Function(std::move(fun)));
